% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_MonoClust.R
\name{as_MonoClust}
\alias{as_MonoClust}
\alias{as_MonoClust.default}
\title{Coerce Similar Object to MonoClust}
\usage{
as_MonoClust(x, ...)

\method{as_MonoClust}{default}(x, ...)
}
\arguments{
\item{x}{An object that can be coerced to MonoClust object.}

\item{...}{For extensibility.}
}
\description{
The function turns a MonoClust-similar object into MonoClust object so it
can use supported functions for MonoClust such as \code{\link[=print.MonoClust]{print.MonoClust()}} and
\code{\link[=plot.MonoClust]{plot.MonoClust()}}.
}
\details{
\code{as_MonoClust()} is an S3 generic. The function itself doesn't run unless
it is implemented for another similar object. Currently, this function is not
implemented within \code{monoClust} package.
}
