% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mono_utils.R
\name{centroid}
\alias{centroid}
\title{Find Centroid of the Cluster}
\usage{
centroid(data, frame, cloc)
}
\arguments{
\item{data}{Original data set.}

\item{frame}{The split tree transferred as data frame.}

\item{cloc}{Vector of current cluster membership.}
}
\value{
A data frame with coordinates of centroids
}
\description{
Centroid is point whose coordinates are the means of their cluster values.
}
\keyword{internal}
