% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{circ_arith}
\alias{circ_arith}
\alias{\%circ+\%}
\alias{\%circ-\%}
\title{Add/Subtract Circular Values in Degrees}
\usage{
x \%circ+\% y

x \%circ-\% y
}
\arguments{
\item{x, y}{Circular values in degrees.}
}
\value{
A value between [0, 360).
}
\description{
Add/subtract two circular variables.
}
\examples{
90 \%circ+\% 90

250 \%circ+\% 200

25 \%circ-\% 80

}
