% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonoClust.R
\name{cshift}
\alias{cshift}
\title{Add/Subtract From A Circular Value}
\usage{
cshift(variable, shift)
}
\arguments{
\item{variable}{circular variable (in degree 0-360).}

\item{shift}{the added value to the circular variable, can be positive or
negative if want to subtract.}
}
\value{
shifted circular value or variable.
}
\description{
Add a value to a circular value or variable. When the value reaches 360 or 0,
it will become 0.
}
