% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.plot.R
\name{cv.plot}
\alias{cv.plot}
\title{Plot the Mean Square Error with error bar for +/- 1 Standard Error}
\usage{
cv.plot(
  cp.table,
  main = "MSE for CV of monothetic clustering",
  ylab = "MSE +/- 1 SE",
  ...
)
}
\arguments{
\item{cp.table}{The output of cv.test or cv.test.mse}

\item{main}{Title of the plot}

\item{ylab}{Text shown on the y-axis}

\item{...}{more args}
}
\value{
A plot
}
\description{
Plot the Mean Square Error with error bar for +/- 1 Standard Error
}
