% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.monoclust.R
\name{error_bar}
\alias{error_bar}
\title{Make Error Bars}
\usage{
error_bar(x, y, upper, lower = upper, length = 0.1, ...)
}
\arguments{
\item{x}{x coordinates.}

\item{y}{y coordinates.}

\item{upper}{Distance from y to the upper bar.}

\item{lower}{Distance from y to the lower bar.}

\item{length}{Length of the horizontal bar.}

\item{...}{Other arguments to \code{\link[graphics:arrows]{graphics::arrows()}}}
}
\value{
Plot
}
\description{
Make Error Bars
}
\keyword{internal}
