% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonoClust.R
\name{find_closest}
\alias{find_closest}
\title{Find the Closest Cut}
\usage{
find_closest(cols)
}
\arguments{
\item{col}{list of quantitative vectors, or a data set}
}
\value{
a quantitative vector
}
\description{
Find the cuts for each quantitative variable. These cuts are what we are
going to consider when thinking about bi-partitioning the data. For a
quantitative column, find the next larger value of each value, if it is the
largest, that value + 1
}
