% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonoClust.R
\name{find_split}
\alias{find_split}
\title{Find the Best Split}
\usage{
find_split(row, data, cuts, dist, variables, weights, minsplit, minbucket)
}
\arguments{
\item{row}{The row in .Cluster_frame that would be assessed.}

\item{data}{Original data set.}

\item{cuts}{Cuts data set, which has the next higher value of each variable
in the original data set.}

\item{dist}{Distance matrix of all observations in the data.}

\item{variables}{List of variables selected for clustering procedure. It
could be a vector of variable indexes, or a vector of variable names.}

\item{weights}{Weights on each observation. Hasn't been implemented in
exported function yet. Vector of 1 for all observations.}

\item{minsplit}{The minimum number of observations that must exist in a node
in order for a split to be attempted.}

\item{minbucket}{The minimum number of observations in any terminal leaf
node. If only one of minbucket or minsplit is specified, the code either
sets minsplit to minbucket*3 or minbucket to minsplit/3, as appropriate.}
}
\value{
This function changes the .Cluster_frame in global environment, it's
not supposed to return anything.
}
\description{
Find the best split in terms of reduction in inertia for the transferred
node, indicate by row. We are keeping all of the information regarding which
clusters we have in. At this point, we want to find the terminal node with
the greatest change in inertia and bi-partition it. The way this is done is
incredibly inefficient. We check every single possible split. This should be
done much better since there is a single local maximum with respect to
inertia so we should use this property to not do an exhaustive check, but to
check the center the 25\% and 75\% and then search more selectively from there.
This is a fairly simple discrete optimization problem and could reduce
computation time .
}
\keyword{internal}
