% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mono_utils.R
\name{get_oper}
\alias{get_oper}
\title{What to Use with ForEach}
\usage{
get_oper(ncores)
}
\arguments{
\item{ncores}{Number of CPU cores on the current host.}
}
\value{
Appropriate operator depending on whether parallel processing is
activated or not.
}
\description{
What to Use with ForEach
}
\keyword{internal}
