% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_MonoClust.R
\name{is_MonoClust}
\alias{is_MonoClust}
\title{Test If The Object is A MonoClust}
\usage{
is_MonoClust(mono_obj)
}
\arguments{
\item{mono_obj}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{MonoClust} class.
}
\description{
This function returns \code{TRUE} for MonoClust, and FALSE for all other objects.
}
