% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.monoclust.R
\name{make_jump_table}
\alias{make_jump_table}
\title{Create Jump Table}
\usage{
make_jump_table(frame)
}
\arguments{
\item{frame}{MonoClust's frame object}
}
\value{
Jump table with \code{number}, \code{var}, and two new columns \code{left} and
\code{right} indicate the left and right number at split.
}
\description{
Create jump table from the MonoClust's frame object. \code{number} and \code{var} will
be used to create the table.
}
\keyword{internal}
