% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mono_utils.R
\name{medoid}
\alias{medoid}
\title{Find Medoid of the Cluster}
\usage{
medoid(members, dist_mat)
}
\arguments{
\item{members}{index vector indicating which observation belongs to the
cluster.}

\item{dist_mat}{distance matrix of the whole data set.}
}
\value{
index of the medoid point in the members vector.
}
\description{
Medoid is the point that has minimum distance to all other points in the
cluster.
}
\keyword{internal}
