% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_node}
\alias{new_node}
\title{Create A New Node for Split Data Frame}
\usage{
new_node(
  number,
  var,
  cut = NA,
  n,
  wt,
  inertia,
  bipartvar = "",
  bipartsplitrow = -99,
  bipartsplitcol = -99,
  inertiadel = 0,
  yval,
  medoid,
  category = 0,
  loc,
  split.order = -99
)
}
\arguments{
\item{number}{Row index of the data frame.}

\item{var}{Whether it is a leaf, or the name of the next split variable.}

\item{cut}{The splitting value, so values (of \code{bipartvar}) smaller than that
go to left branch while values greater than that go to right branch.}

\item{n}{NA}

\item{wt}{NA}

\item{inertia}{Inertia value of the cluster at that node.}

\item{bipartvar}{Name of the next split variable, match \code{var} if \code{var} is not
a leaf.}

\item{bipartsplitrow}{NA}

\item{bipartsplitcol}{NA}

\item{inertiadel}{The proportion of inertia value of the cluster at that node
to the inertia of the root.}

\item{yval}{NA}

\item{medoid}{NA}

\item{category}{Whether splitting variable is categorical (1) or not (0).}

\item{loc}{y-coordinate of the splitting node to facilitate showing on the
tree. See \code{\link[=plot.MonoClust]{plot.MonoClust()}} for details.}

\item{split.order}{Order of the splits. Root is 0, and increasing.}
}
\value{
A tibble with only one row and correct data type for even an
unspecified variables.
}
\description{
This function is just a helper to make sure that the default values of the
split data frame is correct when unspecified. It helps reduce type error,
especially when moving to use dplyr which is stricter in data types.
}
\keyword{internal}
