% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcp.circ.plot.R
\name{pcp.circ.plot}
\alias{pcp.circ.plot}
\title{Plot a parallel coordinate plot with one circular variable}
\usage{
pcp.circ.plot(
  data,
  cir.var = NULL,
  membership = NULL,
  order = NULL,
  col = NULL,
  zero = "east",
  nslabel = TRUE,
  rotation = "counter",
  meds = NULL
)
}
\arguments{
\item{data}{The full data set}

\item{cir.var}{The index or name of the circular variable in the data set}

\item{membership}{A vector of cluster membership for all observations in the
data set}

\item{order}{The order of appearance of the variables in the plot}

\item{col}{List of color for the membership. It should match the number of
groups in membership parameter}

\item{zero}{The position of the zero degree value in the circular variable.
Can be chosen in c('north', 'east', 'west', 'south'). Default is 'east'
side of the circle (usual 0 position in calculus circle)}

\item{nslabel}{Whether to print the North and South label on the circular
variable. Default is TRUE}

\item{rotation}{The positive direction of the circular variable. Default is
'counter'-clockwise. Can be changed to 'clock'.}

\item{meds}{A list of median values of the cluster. The length of this vector
should match the unique values in membership.}
}
\value{
A PCP plot with an ellipse for circular variable
}
\description{
Plot a parallel coordinate plot with one circular variable
}
\examples{
Blank
}
