% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcp.gg.R
\name{pcp.gg}
\alias{pcp.gg}
\title{Parallel Coordinates Plot with Circular Variables}
\usage{
pcp.gg(
  data,
  circ.var,
  order.appear,
  cluster.sol,
  cols,
  is.degree = TRUE,
  shift = 0,
  alpha = 0.1,
  show.medoids = TRUE,
  north = 0,
  cw = FALSE,
  labelsize = 4
)
}
\arguments{
\item{data}{Data set.}

\item{circ.var}{Circular variable(s) in the data set, indicated by names
or index in the data set.}

\item{order.appear}{The order of appearance of the variables, listed by a
vector of names or indices.}

\item{cluster.sol}{A \code{MonoClust} object which contains the cluster results
and memberships.}

\item{cols}{Color of clusters, indicating by a vector. The length of this
vector must be equal to the number of clusters in the cluster.sol object.}

\item{is.degree}{Whether the unit of the circular variables is degree or not
(radian). Default is \code{TRUE}.}

\item{shift}{The shift (offset, rotation) of the circular variable, in
radians. Default is 0 (no rotation).}

\item{alpha}{The transparency of the lines. Default to be 0.1.}

\item{show.medoids}{Whether to highlight the median lines or not. Default is
\code{TRUE}.}

\item{north}{What value of the circular variable is labeled North. Default is
0 radian.}

\item{cw}{Which direction of the circular variable is considered increasing
in value, clockwise (\code{TRUE}) or counter-clockwise (\code{FALSE}). Default is
\code{TRUE}.}

\item{labelsize}{The size of labels on the plot, default is 4.}
}
\value{
A ggplot2 object
}
\description{
Making a parallel coordinates plot with the circular variables are plotted
as ellipses. The function currently works well with data with one circular
variable.
}
\examples{
# Set color constant
COLOR4 <- c("#e41a1c", "#377eb8", "#4daf4a", "#984ea3")
# Reduce the size of the data for for sake of example speed
set.seed(12345)
wind_reduced <- wind_sensit_2007[sample.int(nrow(wind_sensit_2007), 50), ]

sol42007 <- MonoClust(wind_reduced, cir.var = 3, nclusters = 4)

pcp.gg(wind_reduced,
       "WDIR",
       c("WDIR", "has.sensit", "WS"),
       sol42007,
       COLOR4,
       shift = pi*3/4-0.3,
       alpha = 0.1,
       labelsize = 8) +
  theme(panel.background = element_rect(color = NA),
        plot.background = element_rect(color = NA),
        panel.border = element_rect(color = NA),
        panel.grid.major = element_line(color = "#f0f0f0"),
        panel.grid.minor = element_blank(),
        axis.line = element_line(color = "black"),
        legend.key = element_rect(color = NA),
        legend.position = "bottom",
        legend.direction = "horizontal",
        legend.title = element_text(face = "italic"),
        legend.justification = "center",
        text = element_text(size = 24))
}
