% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.rpart}
\alias{plot.rpart}
\title{Plot the splitting tree. Inherited from rpart::plot.rpart}
\usage{
\method{plot}{rpart}(x, uniform = FALSE, branch = 1,
  compress = FALSE, nspace, margin = 0, minbranch = 0.3,
  bar = 0.03, ...)
}
\arguments{
\item{x}{MonoClust object}

\item{uniform}{see Documentation of rpart package}

\item{branch}{see Documentation of rpart package}

\item{compress}{see Documentation of rpart package}

\item{nspace}{see Documentation of rpart package}

\item{margin}{see Documentation of rpart package}

\item{minbranch}{see Documentation of rpart package}

\item{bar}{see Documentation of rpart package}

\item{...}{more args}
}
\value{
Plot of tree
}
\description{
Plot the splitting tree. Inherited from rpart::plot.rpart
}
\examples{
Blank
}
