% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_prep_branch}
\alias{plot_prep_branch}
\title{Calculate Branch Coordinates}
\usage{
plot_prep_branch(x, y, node, branch = 0)
}
\arguments{
\item{x}{Nodes x-coordinates.}

\item{y}{Nodes y-coordinates.}

\item{node}{Nodes row number.}

\item{branch}{Controls the shape of the branches from parent to child node.
Any number from 0 to 1 is allowed. A value of 1 gives square shouldered
branches, a value of 0 give V shaped branches, with other values being
intermediate.}
}
\value{
Branch coordinates in a list of x and y axis.
}
\description{
Calculate Branch Coordinates
}
\keyword{internal}
