% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot the splitting tree. Inherited from rpart::plot_tree}
\usage{
plot_tree(
  x,
  uniform = FALSE,
  branch = 1,
  compress = FALSE,
  nspace,
  margin = 0,
  minbranch = 0.3,
  bar = 0.03,
  rel.loc.x = TRUE,
  ...
)
}
\arguments{
\item{x}{MonoClust object}

\item{uniform}{see Documentation of rpart package}

\item{branch}{see Documentation of rpart package}

\item{compress}{see Documentation of rpart package}

\item{nspace}{see Documentation of rpart package}

\item{margin}{see Documentation of rpart package}

\item{minbranch}{see Documentation of rpart package}

\item{bar}{see Documentation of rpart package}

\item{rel.loc.x}{Whether to use the relative distance between clusters as x
coordinate of the leaves. Default is TRUE.}

\item{...}{more args}
}
\value{
Plot of tree
}
\description{
Plot the splitting tree. Inherited from rpart::plot_tree
}
\keyword{internal}
