% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cv.MonoClust.R
\name{print.cv.MonoClust}
\alias{print.cv.MonoClust}
\title{Print MonoClust Cross-Validation Result}
\usage{
\method{print}{cv.MonoClust}(x, ...)
}
\arguments{
\item{x}{A \code{cv.MonoClust} object (output of \code{\link[=cv.test]{cv.test()}}).}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print MonoClust Cross-Validation Result
}
\examples{
\donttest{
library(cluster)
data(ruspini)

# 10-fold cross-validation
cp_table <- cv.test(ruspini, minnodes = 2, maxnodes = 4)
print(cp_table)
}
}
