% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.MonoClust.R
\name{print.MonoClust}
\alias{print.MonoClust}
\title{Print Monothetic Clustering Results}
\usage{
\method{print}{MonoClust}(x, abbrev = 0, spaces = 2L, digits = getOption("digits"))
}
\arguments{
\item{x}{MonoClust result object.}

\item{abbrev}{Whether print the abbreviated version of labels.}

\item{spaces}{Spaces indent between 2 tree levels.}

\item{digits}{Number of significant digits to print.}
}
\value{
Splitting tree displayed nicely.
}
\description{
Print Monothetic Clustering Results
}
\examples{
library(cluster)
data(ruspini)
ruspini4sol <- MonoClust(ruspini, nclusters = 4)
print(ruspini4sol)
}
