% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonoClust.R
\name{splitter}
\alias{splitter}
\title{Split Function}
\usage{
splitter(data, cuts, split_row, frame, cloc, dist, weights, split_order = 0L)
}
\arguments{
\item{data}{Original data set.}

\item{cuts}{Cuts data set, which has the next higher value of each variable
in the original data set.}

\item{split_row}{The row index in frame that would be split on.}

\item{frame}{The split tree transferred as data frame.}

\item{cloc}{Vector of current cluster membership.}

\item{dist}{Distance matrix of all observations in the data.}

\item{weights}{Weights on each observation. Hasn't been implemented in
exported function yet. Vector of 1 for all observations.}

\item{split_order}{The control argument to see how many split has been done.}
}
\value{
Updated \code{frame} and \code{cloc} saved in a list.
}
\description{
Given the Cluster's frame's row position to split at \code{split_row}, this
function performs the split, calculate all necessary information for the
splitting tree and cluster memberships.
}
\keyword{internal}
