% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.test.R
\name{test_split}
\alias{test_split}
\title{Hypothesis Test at Split}
\usage{
test_split(
  current,
  members,
  members_l,
  members_r,
  auto.pick,
  method,
  data,
  split_var,
  node,
  rep,
  stat
)
}
\arguments{
\item{current}{Row index of the current considered split.}

\item{members, members_l, members_r}{Vector of the index of observations that
are members of the parent node, left child node, and right child node,
respectively.}

\item{auto.pick}{Whether the algorithm stops when p-value becomes larger than
\code{sig.val} or keeps testing and let the researcher pick the final splitting
tree. Default value is \code{FALSE}.}

\item{method}{Can be chosen between 1 (default), 2, or 3. See description
above for the details.}

\item{data}{The data set which is being clustered.}

\item{split_var}{Splitting variable at \code{current} split.}

\item{node}{The current node index}

\item{rep}{Number of permutations to calculate test statistic.}

\item{stat}{Statistic to use, choosing between \code{"F"} (default), \code{"AW"}.}
}
\value{
To be filled
}
\description{
Hypothesis Test at Split
}
\keyword{internal}
