% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{text_impl}
\alias{text_impl}
\title{Implementation of Print Labels on MonoClust Tree}
\usage{
text_impl(
  x,
  splits = TRUE,
  which = 4,
  label = "var",
  FUN = graphics::text,
  all.leaves = FALSE,
  pretty = NULL,
  digits = getOption("digits") - 2,
  tadj = 0.65,
  stats = TRUE,
  use.n = TRUE,
  bars = TRUE,
  legend = FALSE,
  xadj = 1,
  yadj = 1,
  bord = FALSE,
  big.pts = FALSE,
  abbrev = 4,
  cols = NULL,
  rel.loc.x = TRUE,
  ...
)
}
\arguments{
\item{...}{Extra arguments that would be transferred to \code{\link[graphics:text]{graphics::text()}}}
}
\value{
Labels on tree.
}
\description{
This function plots the labels onto the MonoClust tree. It is partially
inspired by rpart package.
}
\keyword{internal}
