% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MonoClust.R
\name{text_MonoClust}
\alias{text_MonoClust}
\title{Title}
\usage{
text_MonoClust(
  x,
  splits = TRUE,
  which = 4,
  label = "var",
  FUN = text,
  all.leaves = FALSE,
  pretty = NULL,
  digits = getOption("digits") - 2,
  tadj = 0.65,
  stats = TRUE,
  use.n = TRUE,
  bars = TRUE,
  legend = FALSE,
  xadj = 1,
  yadj = 1,
  bord = FALSE,
  big.pts = FALSE,
  abbrev = 4,
  cols = NULL,
  rel.loc.x = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
