% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{text_tree}
\alias{text_tree}
\title{Implementation of Print Labels on MonoClust Tree}
\usage{
text_tree(
  x,
  which = 4,
  digits = getOption("digits") - 2,
  stats = TRUE,
  abbrev,
  cols = NULL,
  cols_type = c("l", "p", "b"),
  rel_loc_x = TRUE,
  ...
)
}
\arguments{
\item{x}{MonoClust result object.}

\item{which}{Labeling modes, which are:
\itemize{
\item 1: only splitting variable names are shown, no splitting rules.
\item 2: only splitting rules to the left branches are shown.
\item 3: only splitting rules to the right branches are shown.
\item 4 (default): splitting rules are shown on both sides of branches.
}}

\item{digits}{Number of significant digits to print.}

\item{stats}{Whether to show statistics (cluster sizes and medoid points) on
the tree.}

\item{abbrev}{Whether to print the abbreviated versions of variable names.
Can be either "no" (default), "short", or "abbreviate". Short forms of them
can also be used.

If "no", the labels recorded in \code{x$labels} are used.

If "short", variable names will be turned into "V1", "V2", ...

If "abbreviate", \code{\link[=abbreviate]{abbreviate()}} function will be used. Use the optional
arguments for this function.}

\item{cols}{Whether to shown color bars at leaves or not. It helps matching
this tree plot with other plots whose cluster membership were colored. It
only works when \code{text} is \code{TRUE}. Either \code{NULL}, a vector of one color, or
a vector of colors matching the number of leaves.}

\item{cols_type}{When \code{cols} is set, choose whether the color indicators are
shown in a form of solid lines below the leaves (\code{"l"}), or big points
(\code{"p"}), or both (\code{"b"}).}

\item{rel_loc_x}{Whether to use the relative distance between clusters as x
coordinate of the leaves. Default is TRUE.}

\item{...}{Extra arguments that would be transferred to \code{\link[graphics:text]{graphics::text()}}}
}
\value{
Labels on tree.
}
\description{
This function plots the labels onto the MonoClust tree. It is partially
inspired by rpart package.
}
\keyword{internal}
