% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcp.gg.R
\name{to_deg_rad}
\alias{to_deg_rad}
\alias{torad}
\alias{todeg}
\title{Transform Degree to Radian}
\usage{
torad(x)

todeg(x)
}
\arguments{
\item{x}{A degree value if \code{torad} or radian value if \code{todeg}.}
}
\value{
A radian value if \code{torad} or degree value if \code{todeg}.
}
\description{
This function transforms a circular angle from degree to radian.
}
\examples{
torad(90)

torad(-45)

todeg(pi/2)
}
