% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcp.gg.R
\name{torad}
\alias{torad}
\title{Transform Degree to Radian}
\usage{
torad(x)
}
\arguments{
\item{x}{A degree value.}
}
\value{
A radian value.
}
\description{
Transform Degree to Radian
}
\examples{
torad(90)

torad(45)
}
