% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_depth.R
\name{tree_depth}
\alias{tree_depth}
\title{Find Tree Depth Based on Node Indexes}
\usage{
tree_depth(nodes)
}
\arguments{
\item{nodes}{Vector of node indexes in the tree.}
}
\value{
Depth of the node, with 0 is the root relative to the input.
}
\description{
Find Tree Depth Based on Node Indexes
}
\details{
When building MonoClust tree, the node index was created with the rule that
new node indexes are the split node times 2 plus 0 (left) and 1 (right).
Therefore, this function is just a back-transform, taking a log base 2.
}
