% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.MonoClust.R
\name{tree_walk}
\alias{tree_walk}
\title{Traverse a Tree to Find the Leaves (Terminal Nodes)}
\usage{
tree_walk(new_point, jump_table)
}
\arguments{
\item{new_point}{New data point}

\item{jump_table}{Jump table}
}
\value{
The index of the terminal node after traversing the new data point on
the tree.
}
\description{
Traverse a Tree to Find the Leaves (Terminal Nodes)
}
\keyword{internal}
