% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wind_sensit_2008}
\alias{wind_sensit_2008}
\title{Existence of Microorganisms Carried in Wind}
\format{
A data frame with 673 rows and 3 variables:
\describe{
\item{has.sensit}{A binary variable of the existence of particles in the
wind (1) or not (0).}
\item{WS}{Wind speed measured in m/s.}
\item{WDIR}{Wind direction in degree with 0 indicates "from the north" and
90 degrees indicate "from the east".}
}
}
\source{
Sabacka, M., Priscu, J. C., Basagic, H. J., Fountain, A. G., Wall, D.
H., Virginia, R. A., and Greenwood, M. C. (2012). "Aeolian flux of biotic and
abiotic material in Taylor Valley, Antarctica". In: Geomorphology 155-156,
pp. 102-111. issn: 0169555X. \url{doi:10.1016/j.geomorph.2011.12.009}.
}
\usage{
wind_sensit_2008
}
\description{
Data set is a part of a study on microorganisms carried in strong f\"ohn
winds at the Bonney Riegel location of Taylor Valley, an ice free area in the
Antarctic continent. Wind direction and wind speed data were obtained from
the meteorological station. Wind direction was recorded every 30 seconds and
wind speeds every 4 seconds at 1.15 meters above the ground surface. The
recorded wind directions and speeds were averaged at 15 minute intervals. For
wind direction, as discussed previously, winds from the north are defined as
0/360 degrees and from the east as 90 degrees. 2008 data were collected from
July 7--14, 2008.
}
\keyword{datasets}
