% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{circ_arith}
\alias{circ_arith}
\alias{\%c+\%}
\alias{\%c-\%}
\title{Add/Subtract Circular Values in Degrees}
\usage{
x \%c+\% y

x \%c-\% y
}
\arguments{
\item{x, y}{Circular values in degrees.}
}
\value{
A value between [0, 360).
}
\description{
Add/subtract two circular variables.
}
\examples{
90 \%c+\% 90

250 \%c+\% 200

25 \%c-\% 80

}
