% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonoClust.R
\name{circ_dist}
\alias{circ_dist}
\title{Circular Distance using Gower's}
\usage{
circ_dist(x)
}
\arguments{
\item{x}{a numeric vector of circular values}
}
\value{
object of class "dist"
}
\description{
calculates the distance matrix within a circular variable using Gower's
distance. Written by Garland Will.
}
