% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.test.R
\name{cv.test}
\alias{cv.test}
\title{k-fold Cross validation test on a monothetic clustering result.}
\usage{
cv.test(data, fold = 10, minnodes = 2, maxnodes = 10, ...)
}
\arguments{
\item{data}{data set}

\item{fold}{number of folds (k)}

\item{minnodes}{the minimum number of clusters to be checked}

\item{maxnodes}{the maximum number of clusters to be checked}

\item{...}{}
}
\value{
A list of sum of squares of difference between the predicted and true
values
}
\description{
k-fold Cross validation test on a monothetic clustering result.
}
\examples{
NULL
}
