% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.test (tan.tran@msu.montana.edu).R,
%   R/perm.test.R
\name{F.stat}
\alias{F.stat}
\alias{F.stat}
\title{####################################################
Copy directly from adonis function
Modified to stop at the F.stat
####################################################}
\usage{
F.stat(formula, data = NULL, permutations = 999, method = "bray",
  strata = NULL, contr.unordered = "contr.sum",
  contr.ordered = "contr.poly", parallel = getOption("mc.cores"), ...)

F.stat(formula, data = NULL, permutations = 999, method = "bray",
  strata = NULL, contr.unordered = "contr.sum",
  contr.ordered = "contr.poly", parallel = getOption("mc.cores"), ...)
}
\description{
####################################################
Copy directly from adonis function
Modified to stop at the F.stat
####################################################

####################################################
Copy directly from adonis function
Modified to stop at the F.stat
####################################################
}
