% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_oper}
\alias{get_oper}
\title{What to Use with ForEach}
\usage{
get_oper(x)
}
\arguments{
\item{x}{A binary output of \code{\link[=getDoParWorkers]{getDoParWorkers()}}.}
}
\value{
Appropriate operator depending on whether parallel processing is
activated or not.
}
\description{
What to Use with ForEach
}
\keyword{internal}
