% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoClust-package.R
\docType{package}
\name{monoClust-package}
\alias{monoClust}
\alias{monoClust-package}
\title{monoClust: Perform Monothetic Clustering with Extensions to Circular
    Data}
\description{
Implementation of the Monothetic Clustering
    algorithm (Chavent, 1998 <doi:10.1016/S0167-8655(98)00087-7>) on
    continuous data sets. A lot of extensions are included in the package,
    including applying Monothetic clustering on data sets with circular
    variables, visualizations with the results, and permutation and
    cross-validation based tests to support the decision on the number of
    clusters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vinhtantran/monoClust}
  \item Report bugs at \url{https://github.com/vinhtantran/monoClust/issues}
}

}
\author{
\strong{Maintainer}: Tan Tran \email{vinhtantran@gmail.com} (\href{https://orcid.org/0000-0001-9881-6339}{ORCID})

Authors:
\itemize{
  \item Brian McGuire \email{mcguirebc@gmail.com}
  \item Mark Greenwood \email{greenwood@montana.edu} (\href{https://orcid.org/0000-0001-6933-1201}{ORCID})
}

}
\keyword{internal}
