% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoClust-package.R
\docType{package}
\name{monoClust-package}
\alias{monoClust}
\alias{monoClust-package}
\title{monoClust: Perform Monothetic Clustering with Extensions to Circular}
\description{
Implementation of the Monothetic Clustering algorithm on continuous 
  data set. A lot of extensions are included in the package, including applying
  Monothetic clustering on data set with circular variables, visualizations with
  the results, and permutation and cross-validation based tests to support the
  decision on the number of clusters.
}
\details{
The package performs Monothetic clustering on numerical variables only data.
The extended features include clustering on one circular variable, limiting
the sets of splitting variables. There are also S3 functions such as print
and plot that works for MonoClust output object.
}
\references{
\enumerate{
\item Chavent, M. (1998). A monothetic clustering method. Pattern Recognition
Letters, 19(11), 989–996. https://doi.org/10.1016/S0167-8655(98)00087-7
\item Tran, T. V. (2019). Monothetic Cluster Analysis with Extensions to
Circular and Functional Data. Montana State University - Bozeman.
}
}
\author{
\strong{Maintainer}: Tan Tran \email{vinhtantran@gmail.com} (\href{https://orcid.org/0000-0001-9881-6339}{ORCID})

Authors:
\itemize{
  \item Brian McGuire
  \item Mark Greenwood \email{mark.greenwood@montana.edu}
}

}
\keyword{internal}
