% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonoClust.R
\name{MonoClust}
\alias{MonoClust}
\title{Monothetic Clustering}
\usage{
MonoClust(
  toclust,
  cir.var = NULL,
  variables = NULL,
  distmethod = NULL,
  labels = as.character(seq_len(length(toclust[, 1]))),
  digits = options("digits")$digits,
  nclusters = nrow(toclust),
  minbucket = round(minsplit/3),
  minsplit = 5,
  corders = 2,
  alpha = 0.05,
  perm.test = FALSE,
  ran = 0
)
}
\arguments{
\item{toclust}{data set.}

\item{cir.var}{index or name of the circular variable in the data set.}

\item{variables}{list of variables selected for clustering procedure. It
could be a vector of variable indexes, or a vector of variable names.}

\item{distmethod}{distance method to use with the data set. The default value
is the Euclidean distance for quantitative variables and the Gower's
distance for categorical variables.}

\item{labels}{displayed names of variables.}

\item{digits}{significant number of clusters created}

\item{nclusters}{number of clusters created}

\item{minbucket}{the minimum number of observations in any terminal leaf
node. If only one of minbucket or minsplit is specified, the code either
sets minsplit to minbucket*3 or minbucket to minsplit/3, as appropriate.}

\item{minsplit}{the minimum number of observations that must exist in a node
in order for a split to be attempted.}

\item{corders}{for data set with categorical variables, the number of
dimensions should be used from PCAMix. Taken a value from 1 to 5.}

\item{alpha}{value applied specifically to permutation test.}

\item{perm.test}{whether or not to make a permutation test as stopping
criterion while clustering.}

\item{ran}{this parameter should never be used}
}
\value{
MonoClust object
}
\description{
Creates a MonoClust object after partitioning the data set using Monothetic
Clustering.
}
\examples{
# very simple data set
data(cluster::ruspini)
ruspini4sol <- MonoClust(ruspini, nclusters = 4)
ruspini4sol

# data with categorical variable
data(iris)
iris_cat <- MonoClust(iris, nclusters = 3)
iris_cat

# data with circular variable
data(wind_sensit_2007)

circular_wind <- MonoClust(wind_sensit_2007), cir.var = 3, nclusters = 2)
circular_wind
}
