% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonoClust.R
\name{MonoClust}
\alias{MonoClust}
\title{Title Monothetic Clustering MonoClust creates a MonoClust object after
partitioning the data set using MonoClust clustering.}
\usage{
MonoClust(toclust, cir.var = NULL, variables = NULL, distmethod = NULL,
  labels = as.character(1:length(toclust[, 1])),
  digits = options("digits")$digits, nclusters = nrow(toclust),
  minbucket = round(minsplit/3), minsplit = 5, corders = 2,
  alpha = 0.05, perm.test = FALSE, ran = 0)
}
\arguments{
\item{toclust}{the data set}

\item{cir.var}{the index or name of the circular variable in the data set.
Default is NULL}

\item{variables}{list of variables selected for clustering procedure. It
could be a vector of variable indices, or a vector of variable names.}

\item{distmethod}{distance method to use with the data set. The default value
is the Euclidean distance for quantitative variables and the ... distance
for categorical variables.}

\item{labels}{define names of variables.}

\item{digits}{significant numbers shown in the result.}

\item{nclusters}{number of clusters created.}

\item{minbucket}{the minimum number of observations in any terminal <leaf>
node. If only one of minbucket or minsplit is specified, the code either
sets minsplit to minbucket*3 or minbucket to minsplit/3, as appropriate.}

\item{minsplit}{the minimum number of observations that must exist in a node
in order for a split to be attempted.}

\item{corders}{Blank}

\item{alpha}{This value goes with the permutation test}

\item{perm.test}{Whether or not to make a permutation test while clustering.}

\item{ran}{This parameter should never be used}
}
\description{
Title Monothetic Clustering MonoClust creates a MonoClust object after
partitioning the data set using MonoClust clustering.
}
\examples{
blank
}
