% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonoClust.R
\name{MonoClust}
\alias{MonoClust}
\title{Monothetic Clustering}
\usage{
MonoClust(
  toclust,
  cir.var = NULL,
  variables = NULL,
  distmethod = NULL,
  labels = NULL,
  digits = options("digits")$digits,
  nclusters = 2,
  minsplit = 5,
  minbucket = round(minsplit/3),
  perm.test = FALSE,
  alpha = 0.05,
  ran = 0
)
}
\arguments{
\item{toclust}{Data set as a data frame.}

\item{cir.var}{Index or name of the circular variable in the data set.}

\item{variables}{List of variables selected for clustering procedure. It
could be a vector of variable indexes, or a vector of variable names.}

\item{distmethod}{Distance method to use with the data set. The default value
is the Euclidean distance but Gower will be used if there is circular
variable (\code{cir.var != NULL}).}

\item{labels}{Displayed names of observations. If not specified and data
frame has row names, row names will be used. Otherwise, it is row index.}

\item{digits}{Significant decimal number printed in the output.}

\item{nclusters}{Number of clusters created. Default is 2.}

\item{minsplit}{The minimum number of observations that must exist in a node
in order for a split to be attempted. Default is 5.}

\item{minbucket}{The minimum number of observations in any terminal leaf
node. If not specified, it is set to minsplit/3.#}

\item{perm.test}{Whether or not to make a permutation test as stopping
criterion while clustering. Default is FALSE.}

\item{alpha}{Value applied specifically to permutation test. Only valid when
\code{perm.test = TRUE}.}

\item{ran}{This parameter should never be used.}
}
\value{
MonoClust object, an extension of rpart object.
}
\description{
Creates a MonoClust object after partitioning the data set using Monothetic
Clustering.
}
\examples{
# Very simple data set
library(cluster)
data(ruspini)
ruspini4sol <- MonoClust(ruspini, nclusters = 4)
ruspini4sol

# data with categorical variable
data(iris)
iris_cat <- MonoClust(iris, nclusters = 3)
iris_cat

# data with circular variable
data(wind_sensit_2007)

circular_wind <- MonoClust(wind_sensit_2007, cir.var = 3, nclusters = 2)
circular_wind
}
\references{
\enumerate{
\item Chavent, M. (1998). A monothetic clustering method. Pattern Recognition
Letters, 19(11), 989–996. https://doi.org/10.1016/S0167-8655(98)00087-7
\item Tran, T. V. (2019). Monothetic Cluster Analysis with Extensions to
Circular and Functional Data. Montana State University - Bozeman.
}
}
