% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_node}
\alias{new_node}
\title{Create A New Node for Split Data Frame}
\usage{
new_node(
  number = 1,
  var = "<leaf>",
  cut = NA,
  n = nobs,
  wt = sum(weights[members]),
  inertia = inertia_calc(dismat[members, members]),
  bipartvar = "NA",
  bipartsplitrow = -99,
  bipartsplitcol = -99,
  inertiadel = 0,
  yval = 1,
  medoid = medoid(members, dismat),
  category = 0,
  loc = 0.1,
  split.order = 0
)
}
\arguments{
\item{number}{NA}

\item{var}{NA}

\item{cut}{NA}

\item{n}{NA}

\item{wt}{NA}

\item{inertia}{NA}

\item{bipartvar}{NA}

\item{bipartsplitrow}{NA}

\item{bipartsplitcol}{NA}

\item{inertiadel}{NA}

\item{yval}{NA}

\item{medoid}{NA}

\item{category}{Whether splitting variable is categorical (1) or not (0).}

\item{loc}{NA}

\item{split.order}{NA}
}
\value{
A tibble with only one row and correct data type for even an
unspecified variables.
}
\description{
This function is just a helper to make sure that the default values of the
split data frame is correct when unspecified. It helps reduce type error,
especially when moving to use dplyr which is stricter in data types.
}
\keyword{internal}
