% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MonoClust.R
\name{plot.MonoClust}
\alias{plot.MonoClust}
\title{Plot MonoClust splitting rule tree}
\usage{
\method{plot}{MonoClust}(x, margin, which, abbrev = 4, text = TRUE,
  cols = NULL, ...)
}
\arguments{
\item{x}{MonoClust object}

\item{margin}{Margins to the border of the graphics machine}

\item{which}{Fill in later}

\item{abbrev}{Whether abbreviations of labels are used}

\item{text}{Fill in later}

\item{cols}{Whether to use color or not}

\item{...}{Other args}
}
\value{
Plot of splitting rule
}
\description{
Plot MonoClust splitting rule tree
}
\examples{
NULL
}
