% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.MonoClust.R
\name{predict.MonoClust}
\alias{predict.MonoClust}
\title{Title}
\usage{
\method{predict}{MonoClust}(object, newdata = list(), na.action = na.pass, type = "mean", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
