% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.MonoClust.R
\name{print.MonoClust}
\alias{print.MonoClust}
\title{Print the result of MonoClust clustering}
\usage{
\method{print}{MonoClust}(x, abbrev = 0, spaces = 2, digits = options("digits")$digits, ...)
}
\arguments{
\item{x}{MonoClust result object}

\item{abbrev}{whether print the abbreviated version of labels}

\item{spaces}{spaces between 2 tree levels}

\item{digits}{number of digits shown on the tree}

\item{...}{other args, sending to labels function}
}
\value{
The print result
}
\description{
Print the result of MonoClust clustering
}
\examples{
NULL
}
