% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.MonoClust.R
\name{print.MonoClust}
\alias{print.MonoClust}
\title{Print Monothetic Clustering Results}
\usage{
\method{print}{MonoClust}(x, abbrev = 0, spaces = 2, digits = options("digits")$digits)
}
\arguments{
\item{x}{MonoClust result object.}

\item{abbrev}{Whether print the abbreviated version of labels.}

\item{spaces}{Spaces between 2 tree levels.}

\item{digits}{Number of digits shown on the tree.}
}
\value{
Splitting tree displayed nicely.
}
\description{
Print Monothetic Clustering Results
}
\examples{
library(cluster)
data(ruspini)
ruspini4sol <- MonoClust(ruspini, nclusters = 4)
print(ruspini4sol)
}
