% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.test.R
\name{test_split}
\alias{test_split}
\title{Hypothesis Test at Split}
\usage{
test_split(members_l, members_r, method, data, split_var, rep, stat, ncores)
}
\arguments{
\item{members_l, members_r}{Vector of the index of observations that
are members of the left child node and the right child node, respectively.}

\item{method}{Can be chosen between \code{sw} (simple-withhold, default), \code{rl}
(resplit-limit), or \code{rn} (resplit-nolimit). See Details.}

\item{data}{The data set which is being clustered.}

\item{split_var}{Splitting variable at current split.}

\item{rep}{Number of permutations required to calculate test statistic.}

\item{stat}{Statistic to use. Choosing between \code{"f"} (Calinski-Harabasz's
pseudo-F (Calinski and Harabasz, 1974)) or \code{"aw"} (Average silhoutte width
by Rousseeuw (1987)).}

\item{ncores}{Number of CPU cores on the current host. When set to NULL,
all available cores are used.}
}
\value{
p-value of the test
}
\description{
Hypothesis Test at Split
}
\keyword{internal}
