% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MonoClust.R
\name{text_tree}
\alias{text_tree}
\title{Print Labels on MonoClust Tree}
\usage{
text_tree(x, abbrev, which, rel.loc.x, ...)
}
\arguments{
\item{x}{MonoClust result object.}

\item{abbrev}{Whether to print the abbreviated versions of variable names.
Can be either "no" (default), "short", or "abbreviate". Short forms of them
can also be used.

If "no", the labels recorded in \code{x$labels} are used.

If "short", variable names will be turned into "V1", "V2", ...

If "abbreviate", \code{\link[=abbreviate]{abbreviate()}} function will be used. Use the optional
arguments for this function.}

\item{which}{Fill in later}

\item{rel.loc.x}{Whether to use the relative distance between clusters as x
coordinate of the leaves. Default is TRUE.}

\item{...}{Others.}
}
\value{

}
\description{
Print Labels on MonoClust Tree
}
\keyword{internal}
